feat(compelte profile): moved modal to proper location
